requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-xml==2.0.0
drf-yasg==1.21.3
Flask-MySQLdb==1.0.1
ghp-import==2.1.0
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.3.1
mysql==0.0.3
mysql-connector-python==8.0.30
mysqlclient==2.1.1
packaging==21.3
protobuf==3.20.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
pyyaml_env_tag==0.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
watchdog==2.1.9
Werkzeug==2.2.1
xmltodict==0.13.0
zipp==3.8.1
